requirements.txt
litellm
sf-hamilton[visualization]
hypster
scikit-learn
openpyxl
mlflow